♻️ refactor: localStorage에서 데이터 가져오기 (초기값 설정))
