--- src/main.js
@@ -1,2 +1,0 @@
-import App from "./components/App";
-new App(document.getElementById("app"));
